requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.10.0
click==8.1.3
colorama==0.4.5
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
exceptiongroup==1.0.0
factory-boy==3.2.1
Faker==15.1.1
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
pytest-factoryboy==2.5.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.5
